fix(gpu): change index buffer format from Uint16 to Uint32 for improved compatibility

+++ apps/desktop/src/gpu.rs
@@ -215,7 +215,7 @@ impl GpuState {
                 );
 
                 render_pass
-                    .set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
+                    .set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint32);
 
                 // render_pass.draw(0..self.vertex_count,  0..1);
